Added download billing to post billing file controller/view
