+++ src/EPR.Calculator.Frontend.UnitTests/ViewModels/PostBillingFileViewModelTests.cs
@@ -80,5 +80,65 @@
             // Assert
             Assert.AreEqual(testValue, _testClass.SelectedCalcRunOption);
         }
+
+        [TestMethod]
+        public void CanSetAndGetDownloadResultURL()
+        {
+            // Arrange
+            var fixture = new Fixture().Customize(new AutoMoqCustomization());
+
+            var testValue = fixture.Create<Uri>();
+
+            // Act
+            _testClass.DownloadResultURL = testValue;
+
+            // Assert
+            Assert.AreSame(testValue, _testClass.DownloadResultURL);
+        }
+
+        [TestMethod]
+        public void CanSetAndGetDownloadErrorURL()
+        {
+            // Arrange
+            var fixture = new Fixture().Customize(new AutoMoqCustomization());
+
+            var testValue = fixture.Create<string>();
+
+            // Act
+            _testClass.DownloadErrorURL = testValue;
+
+            // Assert
+            Assert.AreEqual(testValue, _testClass.DownloadErrorURL);
+        }
+
+        [TestMethod]
+        public void CanSetAndGetDownloadTimeout()
+        {
+            // Arrange
+            var fixture = new Fixture().Customize(new AutoMoqCustomization());
+
+            var testValue = fixture.Create<int?>();
+
+            // Act
+            _testClass.DownloadTimeout = testValue;
+
+            // Assert
+            Assert.AreEqual(testValue, _testClass.DownloadTimeout);
+        }
+
+        [TestMethod]
+        public void CanSetAndGetDownloadFileName()
+        {
+            // Arrange
+            var fixture = new Fixture().Customize(new AutoMoqCustomization());
+
+            var testValue = fixture.Create<string>();
+
+            // Act
+            _testClass.DownloadBillingFileName = testValue;
+
+            // Assert
+            Assert.AreEqual(testValue, _testClass.DownloadBillingFileName);
+        }
     }
-}
+  }
