+++ src/EPR.Calculator.Frontend/ViewModels/PostBillingFileViewModel.cs
@@ -13,5 +13,25 @@ namespace EPR.Calculator.Frontend.ViewModels
         public required string SelectedCalcRunOption { get; set; }
 
         public required CalculatorRunStatusUpdateDto CalculatorRunStatus { get; init; }
+
+        /// <summary>
+        /// Gets or sets download Timeout.
+        /// </summary>
+        public int? DownloadTimeout { get; set; }
+
+        /// <summary>
+        /// Gets or sets download result URL.
+        /// </summary>
+        public Uri? DownloadResultURL { get; set; }
+
+        /// <summary>
+        /// Gets or sets download error URL.
+        /// </summary>
+        public string? DownloadErrorURL { get; set; }
+
+        /// <summary>
+        /// Gets or sets download billing file name.
+        /// </summary>
+        public string? DownloadBillingFileName { get; set; }
     }
 }
