+++ src/EPR.Calculator.Frontend/Views/ClassifyRunConfirmation/Index.cshtml
@@ -59,11 +59,6 @@
                                 onclick="downloadFile('@Model.DownloadResultURL', '', '@Model.DownloadErrorURL', event, '@Model.DownloadTimeout', '@HttpContextAccessor.HttpContext?.Session.GetString("accessToken")');">
                             Results file
                         </button>
-                        <button class="govuk-button govuk-button--secondary govuk-!-margin-left-4" data-module="govuk-button"
-                                onclick="downloadFile('@Model.DownloadResultURL', '@Model.DownloadBillingFileName', '@Model.DownloadErrorURL', event, '@Model.DownloadTimeout', '@HttpContextAccessor.HttpContext?.Session.GetString("accessToken")');">
-                            Billing file
-                        </button>
-
                         <hr class="govuk-section-break govuk-section-break--m govuk-section-break--visible">
                         <h2 class="govuk-heading-m">Next step</h2>
                         <p class="govuk-body">Accept all billing instructions contained in the results file.</p>
