+++ src/EPR.Calculator.Frontend/Views/PostBillingFile/Index.cshtml
@@ -1,5 +1,7 @@
 ﻿@using EPR.Calculator.Frontend.ViewModels;
 @using EPR.Calculator.Frontend.Constants;
+@inject IHttpContextAccessor HttpContextAccessor
+
 @model PostBillingFileViewModel;
 @{
     ViewData["Title"] = "Index";
@@ -7,12 +9,12 @@
 }
 <partial name="_BackLinkPartial" model="(@Model.BackLink, @Model.CurrentUser)" />
 <div class="govuk-width-container ">
-        <div class="govuk-grid-row">
-            <div class="govuk-grid-column-two-thirds">
-               @using(Html.BeginForm("","",FormMethod.Post))
-               {
-                    @Html.AntiForgeryToken()
-                    <div class="govuk-form-group">
+    <div class="govuk-grid-row">
+        <div class="govuk-grid-column-two-thirds">
+            @using (Html.BeginForm("", "", FormMethod.Post))
+            {
+                @Html.AntiForgeryToken()
+                <div class="govuk-form-group">
                     <fieldset class="govuk-fieldset">
                         <legend></legend>
                         <span class="govuk-caption-l">@Model.CalculatorRunStatus.CalcName</span>
@@ -57,7 +59,7 @@
                             <button type="submit" class="govuk-button govuk-button--secondary" data-module="govuk-button">
                                 Results file
                             </button>
-                            <button type="submit" class="govuk-button govuk-button--secondary" data-module="govuk-button">
+                            <button type="submit" class="govuk-button govuk-button--secondary" data-module="govuk-button" onclick="downloadFile('@Model.DownloadResultURL', '@Model.DownloadBillingFileName', '@Model.DownloadErrorURL', event, '@Model.DownloadTimeout', '@HttpContextAccessor.HttpContext?.Session.GetString("accessToken")');">
                                 Billing file
                             </button>
                         </div>
@@ -67,8 +69,8 @@
                         </div>
 
                     </fieldset>
-                    </div>
-               }
-            </div>
+                </div>
+            }
         </div>
+    </div>
 </div>
